Dynamic Order Status + Universal Quantity
When an order is created it is written into the Orders table with the "Processing" Status. Over the course of 30 second intervals, the Status then changes to "Shipped" and then to "Delivered". This Status change is also reflected in the View Orders Pop-Up in the account page.

I also made quantity amounts universal. So an item can have a set amount of quantity, and if one unit is bought then the stock quantity is brought down by one. Buyers can only to cart however much stock quantity is available and when when the stock reaches 0 the item card and page says out of stock and prevents buyers from purchasing the product.